refactor: move stale sync stages to constants
