+++ packages/twenty-server/src/modules/calendar/calendar-event-import-manager/constants/calendar-ongoing-stale-sync-stages.constant.ts
@@ -1,0 +1,8 @@
+import { CalendarChannelSyncStage } from 'twenty-shared/types';
+
+export const CALENDAR_ONGOING_STALE_SYNC_STAGES: CalendarChannelSyncStage[] = [
+  CalendarChannelSyncStage.CALENDAR_EVENTS_IMPORT_ONGOING,
+  CalendarChannelSyncStage.CALENDAR_EVENT_LIST_FETCH_ONGOING,
+  CalendarChannelSyncStage.CALENDAR_EVENTS_IMPORT_SCHEDULED,
+  CalendarChannelSyncStage.CALENDAR_EVENT_LIST_FETCH_SCHEDULED,
+];
