+++ packages/twenty-server/src/modules/calendar/calendar-event-import-manager/jobs/calendar-ongoing-stale.job.ts
@@ -10,19 +10,13 @@ import { MessageQueue } from 'src/engine/core-modules/message-queue/message-queu
 import { CalendarChannelEntity } from 'src/engine/metadata-modules/calendar-channel/entities/calendar-channel.entity';
 import { GlobalWorkspaceOrmManager } from 'src/engine/twenty-orm/global-workspace-datasource/global-workspace-orm.manager';
 import { buildSystemAuthContext } from 'src/engine/twenty-orm/utils/build-system-auth-context.util';
+import { CALENDAR_ONGOING_STALE_SYNC_STAGES } from 'src/modules/calendar/calendar-event-import-manager/constants/calendar-ongoing-stale-sync-stages.constant';
 import { isSyncStale } from 'src/modules/calendar/calendar-event-import-manager/utils/is-sync-stale.util';
 import { CalendarChannelSyncStatusService } from 'src/modules/calendar/common/services/calendar-channel-sync-status.service';
 
 export type CalendarOngoingStaleJobData = {
   workspaceId: string;
 };
-
-export const CALENDAR_ONGOING_STALE_SYNC_STAGES: CalendarChannelSyncStage[] = [
-  CalendarChannelSyncStage.CALENDAR_EVENTS_IMPORT_ONGOING,
-  CalendarChannelSyncStage.CALENDAR_EVENT_LIST_FETCH_ONGOING,
-  CalendarChannelSyncStage.CALENDAR_EVENTS_IMPORT_SCHEDULED,
-  CalendarChannelSyncStage.CALENDAR_EVENT_LIST_FETCH_SCHEDULED,
-];
 
 @Processor({
   queueName: MessageQueue.calendarQueue,
